backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic[email]
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart